+++ changelog.md
@@ -1,3 +1,8 @@
+# v1.0.0-beta.3 
+- Added support for mods that add to the right side menu (Texture Loader as an example)
+- Made it "easier" for Android users to move the top levels, (you have to double tap them to enter)
+- Made the "space" button be "enter" instead for selecting levels
+- Fixed issue with battery not being scaled properly when in high quality or low quality
 # v1.0.0-beta.2
 - Fix high texture issue
 - Fix Android crash when pressing the keyback button